apps/backend/profile/requirements.txt
flask
psycopg2-binary
pyjwt
flask-cors
pymysql